Add support for sharing `conftest.py` fixtures
