+++ lib/testing/shared_conftest/__init__.py
@@ -1,0 +1,8 @@
+# NOTE: a lot of these are not intended to be imported and referenced directly
+# and are instead intended to be `*` imported in a `conftest.py` file
+from ._shared_conftest import (
+    event_loop as event_loop,
+    client_fixture as client_fixture,
+    patch_prisma_fixture as patch_prisma_fixture,
+    setup_client_fixture as setup_client_fixture,
+)
